Different statements depending on Gradient Style
---IMPORTANT CHANGE--
Post Processing method arranges Legend differently now for TWO COLOR GRADIENT
The Color Table for the Legend is arranged based on the brightness value of each Colour

This version supports 2 different statement executions depending on the GradientStyle selected.
--IMP---In this version GradientStyle is not set by user yet.

CustomData of Entity is modified inside the PaintGradient()  Method

Arrow color is Re-Applied using a method which consist of a ForLoop
---IMP--- Refactoring needs to be done

At this stage merging this to the main code.
Next stage wll consist of User options and refactoring.