Well Grouped Legend using DataTable with 2 Styles ---IMP--- Bar & Arrows painted in a COMPLETELY Different Way
In this version, the method of creating the Legend is very organized. It uses a DataTable which consists columns for string of Force Range, Start and End Value of Force Range, Colour of Force Range.

This DataTable is crucial. The colours of the DataTable are used to colour the Legend AND
Also, to colour the Bars and Arrows. That is, the Force value of the Bar and Arrow are grouped into the various Force Ranges of the Legend and then they are coloured accordingly.

Hence, ALL force values between 200 and 800 have the same Colour.