Fully Functioning Legend Editor
Fully commented codes

Segregated the Postprocessing Methods into smaller segments. Few of these are public so that they can be accessed by the LegendEditor Class.

LegendDataTable is NO LONGER Re-Initialized. It is simply cleared and re-populated

Add Row functionality disabled because not the Values of the Legend are not custom, they are derived from the value of the ColourTable. Need to find a way around this