Refactored version of the "Color Gradient using Eyeshot" Tasks
The code for the Colour Gradient is completely refactored in this version. The mode of plotting the arrows is reverted back to the old way where they were plotted during the Suspension plotting itself.
This eliminated the need of having 6 seperate methods to plot the force arrows for each corner seperately and hence reducd the lines of the code.