Major Bug Found and resolved Refactored the PreliminaryCheck() Method of the Main Form
In this method there were some GUI operations which I moved to the OuptutGUI() method of the Main Form

I found out that the in the VehicleCADPreprocessor and the EditVehicleCAD methods there was an arguement which accepcted the CAD Usercontrol
In every single call made to these methods only the CADVehicleImputs was being passed. Irrespective of whether the Vehicle had ImportCAD or not.
In this version this bug is resvoled. The XUC_ImportCAD's object is passed where applicable and CADVehicleInputs is passed where there is no ImportCAD condnition