---Optimization of Upright Coordinates Method---
In this method I decided to pass the X/Y/Z coordinates of the Upright Pick Up points as the Genes to the Chromosome to Optimize.

The problem is that, in the Objective Function I need to test that the coordinates generated ALAWAYS satisfy the dimensional constraints of the Upright.

According to Paulo's Suggestion trying out a different method of Optimiziing the Euler Angles