In this version, the Genes are the XYZ Translation Vector, ABG Euler Angles and Inboard Pick Up Points of Toe Link
---IMPORTANT---Error calculation method implemented is WRONG.

The Error calculater in this version is 2 fold and INCORRECT
The first error is the error in Bump Steer which is calculated as deg/mm.
The second is the error in the Link Length which normalized and scaled and Root Mean Squared. Hence unitless.
Hence it is incorrect to club these 2 errors

---IMPORTANT--- In this version to test the Optimizer's ability to produce an orientation which gives 4 link lengths exactly the same and only different I omitted the Bump Steer Error Calcuation.

---IMPORTANT--- The Optimizer is now successfully generating valid positions BUT I DON"T UNDERSTAND THE SIGN CONVENTION FOR ROTATIONS