---Combining Errors of Multiple Objectives---
In this version, the Bump Steer Error was neglected and only the Link Lengths were tried to converge for an Upright Position. I did this because I inteded to solve problems with the link length convergence and intended to use Mutli-Objective Optimization using Pareto Chart to include Bump Steer Optimization

---Understood why the Link lengths were not converging---
It wass because I had not constrained them properly. Initially I had constrained only 3 link lengths so the solver was finding random orientations of the upright. Then I added the Toe Link and it made more sense and finally I added the Pushrod and now the answrs make a lot of sense and seem right.

---Regarding the Multi-Objective Optimization---

No success. Looked into paretto but I didn't understand how to do it.
Normalzing and combining the Bump Steer error doesn't work