---Still Unsuccessfull in getting consistent results--- ---Stoped using Optimizer's Wishbone as a param--- ---Decided to use arbitrary Wishbone Length--- ---Decided to continue RMS error for Wishbone--- ---Decided to continue RMS Error for BS + Wishbone---
It seems that when I allow the Optimizer to select a Wishbone length and trying to solve for it is causing a LOT of fluctiation in the error because the Wishbone length (and hence it's error) is constantly changing.
So I decided to go for a constant and arbitrary wishbone length. Now the solver is trying to solve to attain that wishbone length and hence it is working and in the meantime since the wishbone length corresponds to a particular caster, the caster is also getting solved.