---Some Progresss--- ---Included the Caster error--- ---Solver now reaches reqruired Caster--- ---Link Length Error acceptable--- ---Bump Steer Error not very acceptable. Need to check---
Computed the Error of the Caster in an UNSCALED fashion and returned 0.99 Error in case error value crossed 1 or less than 0. This helped with the Csster. Solver is not able to generate Upright position for the required Caster.

Still need to check if the solutions are repeatable

Bump Steer Error is not very less. Need to think about that.

Link length error max is 0.7mm so i think for now I am going to accept it.