---Added Camber Optimization to the Code--- ---Modified the Double WIshbone Solver--- ---Modified (on temp Basis) the Quadratic Equation Solver---
The DoubleWishboneKinematicSolver was modified a bit. The Wheel Spindle Start, Wheel Spindle End and the COntact Patch are all now calculated using the Top and Bottom Camber Mount Points instead of the UBJ and LBJ.

The QUadraticEquationSolver was modified to accept the value CamberShim Length Chsnge. This is temporary. A more robust way needs to be developed for this