---Implemented an extension in the ExtractOptimizationParams Method---
In the this extension I use the MasterDictionary inside a foreach loop to generatet the Optimized Param.

Created another dictionary of double values called OptimizedParams. Stored all the AdjusmentTools Strings inside the dictionary and double value as 0.

If the Master Dictionary consists of a change requested which requires an adjustment tool then that adjustment tool is found inside the Optimized Params dictionary and populated with the Optimized value