---Fixed a few bug in the SetupChange_Children Form Classs--- ---Beta Testing the algorithm as it works in conjunction with the SetupChage--- ---Adding a method to the AssignVehicleParams Class---
Added 2 methods in the AssignVehicleParams Class.
Now there are 3 methods. One to assign all the Vehicle Params before te Kinematic Solver is Run and one to pass all params after the solver is run. So basically in the new methods the oc_FL[0].scmOP and oc_FL[0].waOP objects are used instead of the vehicle.sc_FL and vehicle.wa_FL objectss

I also needed an additional method to initialize an OutputClass List with 100 objects. This needed to be done because I need to run a motion analysis for Bump Steer Error