---Modifying the Code to see it through to the end---
Created a Class called SetupChange_Outputs to hold all the outputs of the Genetic Algo.

Created a cClass called COnvergance to hold the ffitness value of each of the Setup's

Changed the Assign Fucuntion of the SolverMasterClass to accept resilts of the GA and pass it t the object of the Setupchange_Outputs

Modifying the Results methods to display values from the SetupChange_Outputs objects instead of the SetupChange_ClosedLoopSolver object

Trying to slowly phase out the ClosedLoopSolver Classd.